#fixear todos los ids a Long #ana

+++ src/main/java/com/movies/controller/CategoriaController.java
@@ -87,7 +87,7 @@ public class CategoriaController {
     }
 
     @GetMapping("categorias/delete/{id}")
-    public String deleteCustomer(@PathVariable("id") Integer id) {
+    public String deleteCustomer(@PathVariable("id") Long id) {
         if (!categoriaRepository.existsById(id)) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "categoria no encontrada");
         }
